Maps Angle Grinders (#3593)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
Removes plascutters from:
mudskipper
shetland
rigger
talos
komodo
ranger
replaces plascutters on those vessels with:
angle grinder
bowman headset
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
Using the content that got made specifically for salvaging on ships that
are likely going to salvage....
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl:
balance: Chances are, if there was a plasma cutter on a ship, it's an
angle grinder now. Rejoice. Cry in terror. Whatever honestly.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Co-authored-by: Erika Fox <94164348+Bjarl@users.noreply.github.com>